path-main/requirements.txt
streamlit
numpy
scikit-image
opencv-python-headless
Pillow
